Add location.subtractWithRotation, used for bottom vision. Open up some of the methods and variables for OpenCvVisionProvider so it can be extended more easily. Update ReferenceJobProcessor to use the new subtractWithRotation method on bottom vision. Changed VisionUtils.getPixelCenterOffsets to take double instead of int. This allows vision ops that return subpixel resolution to use that data to full effect. Dumped same method from OpenCvVisionProvider. Don't need two copies of it.

+++ src/main/java/org/openpnp/util/VisionUtils.java
@@ -1,12 +1,10 @@
 package org.openpnp.util;
 
-import java.awt.image.BufferedImage;
-
 import org.openpnp.model.Location;
 import org.openpnp.spi.Camera;
 
 public class VisionUtils {
-    public static Location getPixelCenterOffsets(Camera camera, int x, int y) {
+    public static Location getPixelCenterOffsets(Camera camera, double x, double y) {
         double imageWidth = camera.getWidth();
         double imageHeight = camera.getHeight();
